Segunda parte do trabalho
