+++ main.c
@@ -1,25 +1,7 @@
 #include<stdio.h>
-
-int multiplica (int b1, int b2)
-{
-    int res;
-    res = b1 * b2 ;
-    return(res);
-}
-
-int subtrai (int b1, int b2)
-{
-    int res;
-    res = b1-b2 ;
-    return(res);
-}
-
-int soma (int b1, int b2)
-{
-    int res;
-    res = b1 + b2 ;
-    return(res);
-}
+#include "mult.h"
+#include "sub.h"
+#include "sun.h"
 
 int main(void)
 {
