+++ mult.c
@@ -1,0 +1,9 @@
+#include<stdio.h>
+#include "mult.h"
+
+int multiplica (int b1, int b2)
+{
+    int res;
+    res = b1 * b2 ;
+    return(res);
+}
