+++ mult.h
@@ -1,0 +1,3 @@
+#define __MULT__
+
+int multiplica(int, int);
