+++ sub.c
@@ -1,0 +1,9 @@
+#include<stdio.h>
+#include "sub.h"
+
+int subtrai (int b1, int b2)
+{
+    int res;
+    res = b1-b2 ;
+    return(res);
+}
