+++ sub.h
@@ -1,0 +1,3 @@
+#define __SUB__
+
+int subtrai(int, int);
