+++ sun.c
@@ -1,0 +1,9 @@
+#include<stdio.h>
+#include "sun.h"
+
+int soma (int b1, int b2)
+{
+    int res;
+    res = b1 + b2 ;
+    return(res);
+}
