+++ sun.h
@@ -1,0 +1,3 @@
+#define __SUN__
+
+int soma(int, int);
